Refine code

+++ src/persistence/db_util.js
@@ -24,14 +24,12 @@ export function getDbPath() {
 }
 
 async function copyFile(filename, dbPath, distPath) {
-  try {
-    const src = path.join(dbPath, filename)
-    if (fs.existsSync(src)) {
-      const dist = path.join(distPath, filename)
+  const src = path.join(dbPath, filename)
+  if (fs.existsSync(src)) {
+    const dist = path.join(distPath, filename)
+    if (!fs.existsSync(dist)) {
       fs.writeFileSync(dist, fs.readFileSync(src))
     }
-  } catch (error) {
-    console.log(error)
   }
 }
